Refactor stack params into separate file to prevent import cycle
