+++ App/navigators/CredentialStack.tsx
@@ -6,10 +6,7 @@ import ListCredentials from '../screens/ListCredentials'
 
 import defaultStackOptions from './defaultStackOptions'
 
-export type CredentialStackParams = {
-  Credentials: undefined
-  'Credential Details': { credentialId: string }
-}
+import { CredentialStackParams } from 'types/navigators'
 
 const CredentialStack: React.FC = () => {
   const Stack = createStackNavigator<CredentialStackParams>()
