+++ App/navigators/SettingStack.tsx
@@ -6,10 +6,7 @@ import Settings from '../screens/Settings'
 
 import defaultStackOptions from './defaultStackOptions'
 
-export type SettingsStackParams = {
-  Settings: undefined
-  Language: undefined
-}
+import { SettingsStackParams } from 'types/navigators'
 
 const SettingStack: React.FC = () => {
   const Stack = createStackNavigator<SettingsStackParams>()
