+++ App/screens/CredentialOffer.tsx
@@ -1,6 +1,5 @@
 import type { RouteProp } from '@react-navigation/native'
 import type { StackNavigationProp } from '@react-navigation/stack'
-import type { HomeStackParams } from 'navigators/HomeStack'
 
 import { ConnectionRecord, CredentialRecord, CredentialState } from '@aries-framework/core'
 import { IndyCredentialMetadata } from '@aries-framework/core/build/types'
@@ -14,6 +13,7 @@ import { Colors } from '../Theme'
 import { parseSchema } from '../helpers'
 
 import { Button, ModularView, Label } from 'components'
+import { HomeStackParams } from 'types/navigators'
 
 interface CredentialOfferProps {
   navigation: StackNavigationProp<HomeStackParams, 'Credential Offer'>
